Bug Fixes: Fix bug with both search_results_screen and recipe_screen where it wasnt clearing recipeList state and was refrencing old state

+++ lib/ui/recipe/widgets/dual_recipe_cards_widget.dart
@@ -107,7 +107,7 @@ class _MicInfoCard extends StatelessWidget {
   final int time;
   final List<String> ingredients;
   final int weightWatcher;
-  final int healthScore;
+  final double healthScore;
   final List<NutritionModel> nutrients;
   const _MicInfoCard({
     required this.cardWidth,
